refactor: :art: esporta compomponente per mantenere codice pulito
esporto il componente dei bottoni per le domande
